finished implement of moon_design

+++ lib/src/pages/setting/eh/tagsets/tag_sets_page_logic.dart
@@ -1,3 +1,4 @@
+import 'package:bootstrap_icons/bootstrap_icons.dart';
 import 'package:dio/dio.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
@@ -11,6 +12,7 @@ import 'package:skana_ehentai/src/setting/my_tags_setting.dart';
 import 'package:skana_ehentai/src/setting/user_setting.dart';
 import 'package:skana_ehentai/src/utils/eh_spider_parser.dart';
 import 'package:skana_ehentai/src/utils/toast_util.dart';
+import 'package:skana_ehentai/src/utils/widgetplugin.dart';
 
 import '../../../../database/database.dart';
 import '../../../../exception/eh_site_exception.dart';
@@ -50,9 +52,17 @@ class TagSetsLogic extends GetxController with Scroll2TopLogicMixin {
     state.loadingState = LoadingState.loading;
     updateSafely([tagSetId, bodyId]);
 
-    ({List<({int number, String name})> tagSets, bool tagSetEnable, Color? tagSetBackgroundColor, List<WatchedTag> tags, String apikey}) pageInfo;
+    ({
+      List<({int number, String name})> tagSets,
+      bool tagSetEnable,
+      Color? tagSetBackgroundColor,
+      List<WatchedTag> tags,
+      String apikey
+    }) pageInfo;
     try {
-      pageInfo = await ehRequest.requestMyTagsPage(tagSetNo: state.currentTagSetNo, parser: EHSpiderParser.myTagsPage2TagSetNamesAndTagSetsAndApikey);
+      pageInfo = await ehRequest.requestMyTagsPage(
+          tagSetNo: state.currentTagSetNo,
+          parser: EHSpiderParser.myTagsPage2TagSetNamesAndTagSetsAndApikey);
     } on DioException catch (e) {
       log.error('getTagSetFailed'.tr, e.errorMsg);
       snack('getTagSetFailed'.tr, e.errorMsg ?? '', isShort: true);
@@ -122,7 +132,7 @@ class TagSetsLogic extends GetxController with Scroll2TopLogicMixin {
     }
 
     getCurrentTagSet();
-    
+
     myTagsSetting.refreshOnlineTagSets(state.currentTagSetNo);
   }
 
@@ -148,7 +158,8 @@ class TagSetsLogic extends GetxController with Scroll2TopLogicMixin {
     _updateTag(tagSet);
   }
 
-  Future<void> handleUpdateTagStatus(int tagSetIndex, TagSetStatus newStatus) async {
+  Future<void> handleUpdateTagStatus(
+      int tagSetIndex, TagSetStatus newStatus) async {
     TagSetStatus oldStatus = state.tags[tagSetIndex].watched
         ? TagSetStatus.watched
         : state.tags[tagSetIndex].hidden
@@ -177,7 +188,9 @@ class TagSetsLogic extends GetxController with Scroll2TopLogicMixin {
     updateSafely(['$tagId::${tag.tagId}']);
 
     try {
-      await ehRequest.requestDeleteWatchedTag(watchedTagId: state.tags[tagSetIndex].tagId, tagSetNo: state.currentTagSetNo);
+      await ehRequest.requestDeleteWatchedTag(
+          watchedTagId: state.tags[tagSetIndex].tagId,
+          tagSetNo: state.currentTagSetNo);
     } on DioException catch (e) {
       log.error('deleteTagFailed'.tr, e.errorMsg);
       snack('deleteTagFailed'.tr, e.errorMsg ?? '', isShort: true);
@@ -198,7 +211,8 @@ class TagSetsLogic extends GetxController with Scroll2TopLogicMixin {
       return;
     }
 
-    toast('${'deleteTagSuccess'.tr}: ${state.tags[tagSetIndex].tagData.namespace}:${state.tags[tagSetIndex].tagData.key}');
+    toast(
+        '${'deleteTagSuccess'.tr}: ${state.tags[tagSetIndex].tagData.namespace}:${state.tags[tagSetIndex].tagData.key}');
     state.tags.removeAt(tagSetIndex);
 
     state.updateTagState = LoadingState.idle;
@@ -207,67 +221,58 @@ class TagSetsLogic extends GetxController with Scroll2TopLogicMixin {
     myTagsSetting.refreshOnlineTagSets(state.currentTagSetNo);
   }
 
-  Future<void> showBottomSheet(int index, BuildContext context) async {
+  Future<void> showTrigger(int index, BuildContext context) async {
     Get.focusScope?.unfocus();
 
-    showCupertinoModalPopup(
+    showDialog(
       context: context,
-      builder: (_) => CupertinoActionSheet(
-        actions: <CupertinoActionSheetAction>[
-          CupertinoActionSheetAction(
-            child: Row(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                Icon(Icons.favorite, color: UIConfig.tagSetsPageIconDefaultColor(context)).marginOnly(right: 4),
-                SizedBox(width: 56, child: Text('favorite'.tr)),
-              ],
-            ),
+      builder: (BuildContext context) => moonAlertDialog(
+        context: context,
+        title: "${state.tags[index].tagData.namespace}:${state.tags[index].tagData.key}",
+        columnActions: true,
+        actions: [
+          outlinedButton(
+            isFullWidth: true,
+            onPressed: backRoute,
+            label: 'cancel'.tr,
+          ),
+          filledButton(
+            isFullWidth: true,
+            leading: Transform.translate(offset: Offset(0, 2),child: moonIcon(icon: BootstrapIcons.heart,size: 16),),
+            label: 'favorite'.tr,
             onPressed: () {
               backRoute();
               handleUpdateTagStatus(index, TagSetStatus.watched);
             },
           ),
-          CupertinoActionSheetAction(
-            child: Row(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                Icon(Icons.not_interested, color: UIConfig.tagSetsPageIconDefaultColor(context)).marginOnly(right: 4),
-                SizedBox(width: 56, child: Text('hidden'.tr)),
-              ],
-            ),
+          filledButton(
+            isFullWidth: true,
+            leading: Transform.translate(offset: Offset(0, 2),child: moonIcon(icon: BootstrapIcons.ban,size: 16),),
+            label: 'hidden'.tr,
             onPressed: () {
               backRoute();
               handleUpdateTagStatus(index, TagSetStatus.hidden);
             },
           ),
-          CupertinoActionSheetAction(
-            child: Row(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                Icon(Icons.question_mark, color: UIConfig.tagSetsPageIconDefaultColor(context)),
-                SizedBox(width: 56, child: Text('nope'.tr)),
-              ],
-            ),
+          filledButton(
+            isFullWidth: true,
+            leading: Transform.translate(offset: Offset(0, 1),child: moonIcon(icon: BootstrapIcons.question_lg,size: 16),),
+            label: 'nope'.tr,
             onPressed: () {
               backRoute();
               handleUpdateTagStatus(index, TagSetStatus.nope);
             },
           ),
-          CupertinoActionSheetAction(
-            child: Row(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                Icon(Icons.delete, color: UIConfig.alertColor(context)),
-                SizedBox(width: 56, child: Text('delete'.tr)),
-              ],
-            ),
+          filledButton(
+            isFullWidth: true,
+            label: 'delete'.tr,
+            color: UIConfig.alertColor(context),
             onPressed: () {
               backRoute();
               deleteTag(index);
             },
           ),
         ],
-        cancelButton: CupertinoActionSheetAction(child: Text('cancel'.tr), onPressed: backRoute),
       ),
     );
   }
@@ -308,7 +313,8 @@ class TagSetsLogic extends GetxController with Scroll2TopLogicMixin {
       return;
     }
 
-    int tagIndex = state.tags.indexWhere((element) => element.tagId == tag.tagId);
+    int tagIndex =
+        state.tags.indexWhere((element) => element.tagId == tag.tagId);
     state.tags[tagIndex] = tag;
     state.updateTagState = LoadingState.idle;
 
@@ -321,7 +327,8 @@ class TagSetsLogic extends GetxController with Scroll2TopLogicMixin {
   Future<void> _translateTagNamesIfNeeded() async {
     if (tagTranslationService.isReady) {
       for (WatchedTag tagSet in state.tags) {
-        TagData? tagData = await tagTranslationService.getTagTranslation(tagSet.tagData.namespace, tagSet.tagData.key);
+        TagData? tagData = await tagTranslationService.getTagTranslation(
+            tagSet.tagData.namespace, tagSet.tagData.key);
         if (tagData != null) {
           tagSet.tagData = tagData;
         }
